Clean up remaining feedback/bug-reporting items.
- Extract GlobalFeedbackDialog from __root.tsx into its own component.
- Keep the Bug category chip selected through the dialog close transition
  by only syncing category state while the dialog is open.
- Deduplicate sidebar bug-report entry points with a single openBugReport
  callback and add it to the search palette actions memo deps.
- Live browser E2E verified the bug button opens the feedback dialog with
  Bug preselected.

All checks pass: bun fmt, bun lint, bun typecheck, focused Vitest suite,
and full web test suite (run with 10s test timeout to avoid a slow-load
flake on this machine).
