+++ apps/web/src/components/Sidebar.tsx
@@ -1591,6 +1591,10 @@ export default function Sidebar() {
   const [createProjectDialogOpen, setCreateProjectDialogOpen] = useState(false);
   const [searchPaletteOpen, setSearchPaletteOpen] = useState(false);
   const openFeedbackDialog = useFeedbackDialogStore((state) => state.openDialog);
+  const openBugReport = useCallback(
+    () => openFeedbackDialog(undefined, "bug"),
+    [openFeedbackDialog],
+  );
   const [searchPaletteMode, setSearchPaletteMode] = useState<SidebarSearchPaletteMode>("search");
   const projectAdditionLockRef = useRef(false);
   const [renameDialogThreadId, setRenameDialogThreadId] = useState<ThreadId | null>(null);
@@ -5604,7 +5608,7 @@ export default function Sidebar() {
         label: "Report a bug",
         description: "Open the bug report dialog with the Bug category selected.",
         keywords: ["bug", "report", "issue", "problem", "broken", "crash", "synara"],
-        run: () => openFeedbackDialog(undefined, "bug"),
+        run: openBugReport,
       },
       {
         id: "settings",
@@ -5668,6 +5672,7 @@ export default function Sidebar() {
       importThreadShortcutLabel,
       newChatShortcutLabel,
       newThreadShortcutLabel,
+      openBugReport,
       openSpaceCreator,
       spaces,
       usageSettingsShortcutLabel,
@@ -6458,7 +6463,7 @@ export default function Sidebar() {
                   icon={BugIcon}
                   label="Report a bug"
                   tooltip="Report a bug"
-                  onClick={() => openFeedbackDialog(undefined, "bug")}
+                  onClick={openBugReport}
                 />
                 {showDesktopUpdateButton ? (
                   <Tooltip>
