+++ apps/web/src/routes/__root.tsx
@@ -25,16 +25,12 @@ import { QueryClient, useQuery, useQueryClient } from "@tanstack/react-query";
 import { Throttler } from "@tanstack/react-pacer";
 
 import { APP_DISPLAY_NAME, APP_VERSION } from "../branding";
-import { buildFeedbackSubmission } from "../feedback";
-import { buildBugReportDiagnostics, buildGithubIssueInterviewPrompt } from "../feedbackGithubIssue";
-import { useHandleNewThread } from "../hooks/useHandleNewThread";
-import { appendComposerPromptText } from "../lib/chatReferences";
 import { DesktopWindowControls } from "../components/DesktopWindowControls";
 import { RunningChatsQuitCoordinator } from "../components/RunningChatsQuitCoordinator";
 import { AppSnapCoordinator } from "../components/AppSnapCoordinator";
 import { AppSnapWelcomeDialog } from "../components/AppSnapWelcomeDialog";
 import { QueuedComposerDrainCoordinator } from "../components/QueuedComposerDrainCoordinator";
-import { FeedbackDialog } from "../components/FeedbackDialog";
+import { GlobalFeedbackDialog } from "../components/GlobalFeedbackDialog";
 import { SETTINGS_TARGETS } from "../settingsNavigation";
 import ShortcutsDialog from "../components/ShortcutsDialog";
 import WhatsNewDialog from "../components/WhatsNewDialog";
@@ -47,8 +43,6 @@ import { useGitProgressToastPreview } from "../components/useGitProgressToastPre
 import { resolveAndPersistPreferredEditor } from "../editorPreferences";
 import { useFeatureFlags } from "../featureFlags";
 import { useFocusedChatContext } from "../focusedChatContext";
-import { useFeedbackDialogStore } from "../feedbackDialogStore";
-import type { FeedbackThreadContext } from "../feedback";
 import { isTerminalFocused } from "../lib/terminalFocus";
 import {
   invalidateProviderUsageQueries,
@@ -750,69 +744,6 @@ function GlobalShortcutsDialog() {
         terminalOpen,
         terminalWorkspaceOpen,
       }}
-    />
-  );
-}
-
-function GlobalFeedbackDialog() {
-  const { activeProject, activeProjectId, activeThread } = useFocusedChatContext();
-  const isOpen = useFeedbackDialogStore((state) => state.isOpen);
-  const requestedContext = useFeedbackDialogStore((state) => state.context);
-  const requestedInitialCategory = useFeedbackDialogStore((state) => state.initialCategory);
-  const setOpen = useFeedbackDialogStore((state) => state.setOpen);
-
-  const { handleNewThread, projects } = useHandleNewThread();
-  const hasProjects = projects.length > 0;
-
-  const context: FeedbackThreadContext = requestedContext ?? {
-    provider: activeThread?.modelSelection.provider ?? null,
-    model: activeThread?.modelSelection.model ?? null,
-    projectKind: activeProject?.kind ?? null,
-    environmentMode: activeThread?.envMode ?? null,
-    runtimeMode: activeThread?.runtimeMode ?? null,
-    interactionMode: activeThread?.interactionMode ?? null,
-    sessionStatus: activeThread?.session?.status ?? null,
-    latestTurnState: activeThread?.latestTurn?.state ?? null,
-    messageCount: activeThread?.messages.length ?? 0,
-    activityCount: activeThread?.activities.length ?? 0,
-    hasPendingApproval: activeThread?.hasPendingApprovals === true,
-    hasPendingUserInput: activeThread?.hasPendingUserInput === true,
-    hasThreadError: Boolean(activeThread?.error),
-  };
-
-  const onDraftGithubIssue = async (details: string) => {
-    const projectId = activeProjectId ?? projects[0]?.id ?? null;
-    if (!projectId) {
-      throw new Error("No project available.");
-    }
-
-    const submission = buildFeedbackSubmission({ category: "bug", details, context });
-    const prompt = buildGithubIssueInterviewPrompt({
-      details: submission.details,
-      diagnosticsSummary: buildBugReportDiagnostics(submission.diagnostics),
-    });
-
-    const threadId = await handleNewThread(projectId, { fresh: true });
-    if (!threadId) {
-      throw new Error("Could not open a draft thread.");
-    }
-
-    appendComposerPromptText(threadId, prompt);
-    setOpen(false);
-    toastManager.add({
-      type: "success",
-      title: "Bug-report thread ready",
-      description: "Review the prompt and send it.",
-    });
-  };
-
-  return (
-    <FeedbackDialog
-      open={isOpen}
-      context={context}
-      initialCategory={requestedInitialCategory}
-      onOpenChange={setOpen}
-      onDraftGithubIssue={hasProjects ? onDraftGithubIssue : undefined}
     />
   );
 }
